Add tabs to the drawer and decompose the lists list.
